pyproject.toml
[project]
name = "financialtools"
version = "0.1.0"
description = "A package for processing financial data"
authors = [{ name = "Luigi", email = "luigi@example.com" }]
readme = "README.md"
dependencies = [
    "pandas>=1.5",
    "numpy>=1.23",
    "requests>=2.28",
    "yfinance>=0.2",
    "langchain>=1.0,<2.0",
    "langchain-core>=1.0,<2.0",
    "langchain-openai>=0.2",
    "langgraph>=1.0,<2.0",
    "langsmith>=0.3.0",
    "python-dotenv>=1.0"
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
# Explicitly include only your package folder
packages = ["financialtools"]

[tool.setuptools.package-data]
# If you have specific non-python files inside 'financialtools' to include
# "*" = ["*.json", "*.txt"]